Destructibles
Blueprints for all destructibles so that when they are done with crumbling they can be deleted from the scene. This prevents memory leaks while running the game.
Added a tag to NPC and scripting in the Player blueprint to decrease score if an NPC is hit.

All destructibles have a value of 1 currently. If you want to change the values go ahead. I will post a .txt file in discord with what asset belongs to what blueprint